% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{.dr_calc_date}
\alias{.dr_calc_date}
\title{Calculate date based on \code{Year}, \code{Month}, and \code{Day}.}
\usage{
.dr_calc_date(d)
}
\arguments{
\item{d}{A DuckDB connection table or a dataframe containing at least the columns:
\code{Year} (numeric or integer), \code{Month} (numeric or integer), and \code{Day} (numeric or integer).}
}
\value{
The input table (DuckDB table or dataframe) with an additional \code{date} column,
or \code{NULL} if the input is neither a DuckDB table nor a dataframe.
}
\description{
This function adds a new column \code{date} to the input table (DuckDB table or dataframe),
computed from the \code{Year}, \code{Month}, and \code{Day} columns. If the provided input is neither
a DuckDB table nor a dataframe, the function returns \code{NULL} and informs the user.
}
\details{
The function checks if the required columns \code{Year}, \code{Month}, and \code{Day} exist,
and throws an error if they are missing.
}
