% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_convertions.R
\name{.dr_calc_n}
\alias{.dr_calc_n}
\title{Calculate column \code{n} based on \code{DataType} and \code{HaulDur}.}
\usage{
.dr_calc_n(hh, hl)
}
\arguments{
\item{hh}{A table containing columns \code{.id}, \code{DataType}, and \code{HaulDur}.}

\item{hl}{A table containing columns \code{.id} and \code{HLNoAtLngt}.}
}
\value{
A table created by joining \code{hh} with \code{hl}, with an additional column \code{n}:
\itemize{
\item If \code{DataType == "R"}, \code{n} is set to \code{HLNoAtLngt} (data by haul).
\item If \code{DataType == "C"}, \code{n} is calculated as \code{HLNoAtLngt * HaulDur / 60} (CPUE, number per hour).
\item If \code{DataType == "P"} or \code{"S"}, \code{n} is set to \code{-9999} (unsupported case).
\item If \code{DataType == "-9"} or \code{is.na(DataType)}, \code{n} is set to \code{NA}.
\item All other \code{DataType} values default to \code{-10000} (unexpected).
}
}
\description{
This function calculates a new column \code{n} for a table (\code{hl}) by joining it with another table (\code{hh}).
The calculation of \code{n} depends on the values of the \code{DataType} column and the duration of the haul (\code{HaulDur}).
The function ensures all required columns exist in the input tables before processing.
}
