% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{.dr_length_cm}
\alias{.dr_length_cm}
\title{Standardize length to cm}
\usage{
.dr_length_cm(d)
}
\arguments{
\item{d}{A dataframe containing at least two columns: \code{LngtCode} (character) and \code{LngtClass} (numeric).}
}
\value{
The input table with an additional column \code{length}.
}
\description{
Function adds a new column \code{length} to the input dataframe, computed based on the \code{LngtCode} and \code{LngtClass} values.
}
\details{
The transformation rules for the \code{length} column are as follows:
\itemize{
\item \code{LngtCode == "-9"}: The \code{length} is set to \code{NA} (assumed to have been handled upstream).
\item \code{LngtCode \%in\% c(".", "0")}: The \code{length} is computed as the floor of \code{LngtClass} divided by 10.
\item \code{LngtCode \%in\% c("1", "2", "5")}: The \code{length} is directly set as \code{LngtClass}.
\item Otherwise: \code{length} is set to \code{NA}.
}
}
