% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{.dr_n_and_cpue}
\alias{.dr_n_and_cpue}
\title{Numbers caught and the CPUE in each length class}
\usage{
.dr_n_and_cpue(d)
}
\arguments{
\item{d}{DATRAS length table (HL) containing columns \code{.id}, \code{HLNoAtLngt}, \code{.datatype}, and \code{.effort}.}
}
\value{
The input table with additional columns \code{n} and \code{cpue}.
}
\description{
This function generates two new column \code{n} (numbers caught) and
\code{cpue} (numbers caught per one hour towing) for the length table (HL).
}
