% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{dr_add_length_mm}
\alias{dr_add_length_mm}
\title{Add \code{length_mm}}
\usage{
dr_add_length_mm(d)
}
\arguments{
\item{d}{A data frame or lazy table with \code{LengthCode} and
\code{LengthClass}.}
}
\value{
\code{d} with an added integer \code{length_mm} column.
}
\description{
Converts \code{LengthClass} to millimetres using \code{LengthCode}.
}
\seealso{
\code{\link{dr_add_length_cm}}
}
