% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{dr_add_starttime}
\alias{dr_add_starttime}
\title{Calculate timestamp based on \code{Year}, \code{Month}, \code{Day}, and \code{StartTime}/\code{TimeShot}.}
\usage{
dr_add_starttime(d)
}
\arguments{
\item{d}{A DuckDB connection table or a dataframe containing at least the columns:
\code{Year} (numeric or integer), \code{Month} (numeric or integer), \code{Day} (numeric or integer),
and \code{StartTime}/\code{TimeShot} (numeric or character in \code{HHMM} format).}
}
\value{
The input table (DuckDB table or dataframe) with an additional \code{timestamp} column,
calculated as a POSIXct timestamp, or \code{NULL} if the input is neither a DuckDB table nor a dataframe.
}
\description{
This function calculates a precise timestamp using the \code{Year}, \code{Month}, \code{Day}, and \code{StartTime} or \code{TimeShot} columns
from the input table (DuckDB table or dataframe). The time column is expected to be a numeric
or character string representing time in 24-hour \code{HHMM} format. It ensures all required columns are present
in the input and adjusts behavior based on whether the input is a DuckDB table or a standard dataframe.
}
