% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_products.R
\name{dr_catch_weight}
\alias{dr_catch_weight}
\title{Get catch weights}
\usage{
dr_catch_weight(latin, not_numbers = TRUE)
}
\arguments{
\item{latin}{Latin species name}

\item{not_numbers}{A boolean (default TRUE), indicating if total abundance
(TotalNo) should also be calculated}
}
\value{
A DuckDB view
}
\description{
Calculate the total reported catch numbers and weight by species and haul.
}
\details{
The functions is supposed to give the same results as icesDatras::getCatchWgt.
If argument "not_numbers" is set to FALSE this may though not hold true.
}
