% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr_coastline}
\alias{dr_coastline}
\title{A valid aphia id - latin name lookup}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 102 rows and 1 columns.
}
\source{
R-package rnaturalearth
}
\usage{
dr_coastline
}
\description{
Simple shapefile
}
\keyword{datasets}
