% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_products.R
\name{dr_con_by_haul}
\alias{dr_con_by_haul}
\title{Get numbers and weights summary of species by haul}
\usage{
dr_con_by_haul(trim = TRUE)
}
\arguments{
\item{trim}{Boolean (default TRUE), controls if additional non-essential variables are returned or not.}
}
\value{
A summarized \code{data.frame} with the following columns:
\itemize{
\item \code{.id}: Unique haul identifier.
\item \code{latin}: Latin species name, identifying the species.
\item \code{n_haul}: Actual number of individuals per haul.
\item \code{w_haul}: Actual weight (kg) per haul.
\item \code{n_hour}: Number of individuals raised to a 1-hour haul.
\item \code{w_hour}: Weight (kg) raised to a 1-hour haul.
}

If trim is FALSE then additional variables return
\itemize{
\item \code{TotalNumber}: Total number of individuals of the species caught - nonstandardized.
\item \code{SpeciesCategoryWeight}: Total weight (kg) of the species caught - nonstandardized.
\item \code{DataType}: Type of data recording, if value is "C", SpeciesCategoryWeight is in unit per 60 minute hauling, otherwise unit is in reported haul ducation.
\item \code{HaulDuration}: Type of data collected, e.g., raised to 1 hour or raw.
}
}
\description{
This function calculates and summarizes the total numbers and weights of species reported in a survey.
By processing fishing haul-level (HH) and catch-level (HL) data, it generates the aggregate number (\code{n_haul}) and
weight (\code{w_haul}) of species caught per haul, alongside estimates standardized to a 1-hour haul duration (\code{n_hour}, \code{w_hour}).
}
\details{
The final output is a summary by haul identifiers and species, enabling an overview of catch numbers
and weights. Additionally, the function aims to provide similar weight calculations as \code{icesDatras::getCatchWgt},
making its results comparable to external references.
}
\note{
".id" are variables Survey, Year, Quarter, Country, Platform, Gear, StationName and HaulNumber catenated, separated by ":".
}
\seealso{
\code{icesDatras::getCatchWgt} for an alternative approach to computing total catch weight by species and haul.

\code{\link{dr_con}} for information about connecting to DuckDB tables.
}
