% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_connection.R
\name{dr_con_species}
\alias{dr_con_species}
\title{Connect to the Species WoRMS Dataset}
\usage{
dr_con_species()
}
\value{
A \code{duckdb} object connected to the parquet file.
}
\description{
This function provides a connection to the 'species_worms' dataset stored in a Parquet file.
The dataset was created based on the ICES vocabulary code list 'SpecWoRMS'.
}
\details{
The parquet file was generated from the ICES vocabulary code list "SpecWoRMS" with the following steps:
\itemize{
\item The code list is obtained using \code{icesVocab::getCodeList("SpecWoRMS")}.
\item It is processed to contain two columns: \code{Valid_Aphia} (integer) and \code{latin} (scientific name).
\item The resulting data is written as a Parquet file
}

Path to the Parquet file: \url{https://heima.hafro.is/~einarhj/datras/species_worms.parquet}
}
\examples{
# Example of connecting to the species dataset
species <- dr_con_species()
dplyr::glimpse(species)  # Peek at the dataset

}
