% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_download.R
\name{dr_download}
\alias{dr_download}
\title{Download DATRAS parquet files}
\usage{
dr_download(
  recordtype = c("HH", "HL", "CA"),
  url = "https://heima.hafro.is/~einarhj/datras",
  dest_directory = "data"
)
}
\arguments{
\item{recordtype}{Character. Base directory where parquet files will be written.}

\item{url}{path, default "https://heima.hafro.is/~einarhj/datras"}

\item{dest_directory}{path local storage directory (default "data")}
}
\value{
Invisibly returns \code{NULL}. Called for its side effects.
}
\description{
Downloads HH, HL, and CA parquet files
}
