% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_exchange.R
\name{dr_download_data}
\alias{dr_download_data}
\title{Download DATRAS tables}
\usage{
dr_download_data(
  surveys = NULL,
  years = NULL,
  quarters = NULL,
  outpath = "data",
  filetype = "parquet"
)
}
\arguments{
\item{surveys}{Survey to get, if none specified get all}

\item{years}{Years to download, if none specified get all from 1990 onwards. Not yet active.}

\item{quarters}{Quarters to download, if none specified get all from 1990 onwards. Not yet active.}

\item{outpath}{The path (default 'data') where saved DATRAS exchange files are stored}

\item{filetype}{File type (default 'parquet'). Currently inactive.
'}
}
\value{
Files on disk
}
\description{
Each DATRAS table (HH, HL, CA and FL) is downloaded and saved.
}
