% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_latin.R
\name{dr_download_data_latin}
\alias{dr_download_data_latin}
\title{Download DATRAS latin tables}
\usage{
dr_download_data_latin(
  surveys = NULL,
  years = NULL,
  quarters = NULL,
  outpath = "data",
  filetype = "parquet"
)
}
\arguments{
\item{surveys}{Survey to get, if none specified get all}

\item{years}{Years to download, if none specified attempt to download all}

\item{quarters}{Quarters to download, if none specified attempt to download all}

\item{outpath}{The path (default 'data') where saved DATRAS exchange files are stored}

\item{filetype}{File type (default 'parquet'). Currently inactive.
'}
}
\value{
Files on disk
}
\description{
Each DATRAS latin table (HH, HL, CA) is downloaded and saved.
}
