% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr_fields}
\alias{dr_fields}
\title{Datras variable types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 292 rows and 5 columns.
}
\source{
\url{https://datras.ices.dk/WebServices/DATRASWebService.asmx/getDatrasFieldList} and then some
}
\usage{
dr_fields
}
\description{
A table containing the variable (field) types of the datras data
}
\details{
A data frame with 208 rows and 5 columns:
\describe{
\item{RecordHeader}{The DATRAS data type - "HH": haul data, "HL": length-based data, "CA": age-based data}
\item{FieldName}{DATRAS variable name as returned by icesDatras::get_datras_unaggregated_data}
\item{FieldNameOld}{DATRAS variable name as returned by icesDatras::getDatras}
\item{DataFormat}{The value type, char (character), int (ingeger) and decimal (numeric)}
\item{Description}{Some description}
}
}
\keyword{datasets}
