% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_get_specs.R
\name{dr_get_specs}
\alias{dr_get_specs}
\title{Get DATRAS Field Specifications}
\usage{
dr_get_specs(
  url = "https://datras.ices.dk/WebServices/DATRASWebService.asmx/getDatrasFieldList"
)
}
\arguments{
\item{url}{A character string specifying the URL of the DATRAS web service.
Defaults to \code{"https://datras.ices.dk/WebServices/DATRASWebService.asmx/getDatrasFieldList"}.}
}
\value{
A tibble (data frame) containing the field specifications with the
following columns:
\itemize{
\item \code{RecordHeader}: A character string indicating the record header type.
\item \code{FieldName}: A character string specifying the field name.
\item \code{FieldNameOld}: A character string specifying the old field name.
\item \code{DataFormat}: A character string representing the format of the data.
\item \code{Description}: A character string describing the field.
}
}
\description{
This function fetches XML data from the DATRAS web service, parses the content,
and returns the field specifications as a data frame.
}
\examples{
# Fetch the default DATRAS field specifications:
specs <- dr_get_specs()

# View the first few rows:
head(specs)

}
