% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_exchange.R
\name{dr_getFL}
\alias{dr_getFL}
\title{Retrieve DATRAS Flex (FL) Table}
\usage{
dr_getFL(surveys, years, quarters, quiet = TRUE)
}
\arguments{
\item{surveys}{A character string, the survey acronym (e.g., \code{"NS-IBTS"}).}

\item{years}{An integer vector of years (e.g., \code{2010} or \code{2005:2010}).}

\item{quarters}{An integer vector of quarters (1, 2, 3, or 4).}

\item{quiet}{Logical; if \code{TRUE} (default), suppress messages.}
}
\value{
A tibble with station data or an empty tibble if no data is available.
}
\description{
Retrieve DATRAS flex (FL) data for a given survey, year(s), and quarter(s).
This function is a wrapper around \code{icesDatras::getDATRAS(record = "FL", ...)} that ensures variable types are properly set.
}
