% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr_latin_aphia}
\alias{dr_latin_aphia}
\title{Coastline covering majority of DATRAS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2011 rows and 3 columns.
}
\source{
R-package worrms
}
\usage{
dr_latin_aphia
}
\description{
A data frame with over 2000 rows and 3 columns:
\describe{
\item{Valid_Aphia}{Species id in the DATRAS exchange data}
\item{latin}{Species latin name}
\item{species}{Species english name}
}
}
\keyword{datasets}
