% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr_lookup_species}
\alias{dr_lookup_species}
\title{A table of english and latin species names and aphia}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2019 rows and 3 columns.
}
\source{
R-package worrms and aphia in HL- and CA-tables
}
\usage{
dr_lookup_species
}
\description{
A data frame with over 2000 rows and 3 columns:
\describe{
\item{aphia}{Species id in the DATRAS exchange data}
\item{latin}{Species latin name}
\item{species}{Species english name}
}
}
\keyword{datasets}
