% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_tidy.R
\name{dr_tidyhl}
\alias{dr_tidyhl}
\title{Tidy DATRAS length (HL) data}
\usage{
dr_tidyhl(d)
}
\arguments{
\item{d}{A duckdb-table connection or a tibble containing HL data.}
}
\value{
A duckdb-table connection or a tibble with tidied HL data.
}
\description{
Cleans and formats length (HL) data, removing records with missing length class or numbers, converting length units, applying subfactors, and finalizing aphia codes.
}
