% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arni_old_suggetions.R
\name{getIndicesAllYears2}
\alias{getIndicesAllYears2}
\title{Get Survey Indices for All Years}
\usage{
getIndicesAllYears2(survey, quarter, species, quiet = FALSE)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{quarter}{the quarter of the year the survey took place, i.e. 1, 2, 3 or 4.}

\item{species}{the Aphia species code for the species of interest.}

\item{quiet}{Boolean (default FALSE)}
}
\value{
A data frame containing Year in the first column and ages in subsequent columns.
}
\description{
Get age-based indices of abundance by species and survey.
}
\examples{
\dontrun{
haddock_aphia <- icesVocab::findAphia("haddock")
index <- getIndicesAllYears2(survey = "NS-IBTS", quarter = 3, species = haddock_aphia)
str(index)
}

}
