% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_miscallenous.R
\name{dr_add_id}
\alias{dr_add_id}
\title{Generate a haul id}
\usage{
dr_add_id(d)
}
\arguments{
\item{d}{A DATRAS table, any of HH, HL or CA}
}
\value{
A table with an additional variable .id
}
\description{
Generate a haul id
}
