% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_transformation.R
\name{dr_add_length_cm}
\alias{dr_add_length_cm}
\title{Add a standardized \code{length_cm} column to the input table}
\usage{
dr_add_length_cm(d, LengthCode = LengthCode, LengthClass = LengthClass)
}
\arguments{
\item{d}{A dataframe or DuckDB table containing at least two columns:
a column for \code{LengthCode} and a column for \code{LengthClass}.}

\item{LengthCode}{The column specifying the length code (unquoted).
Defaults to \code{LengthCode}.}

\item{LengthClass}{The column specifying the length class (unquoted).
Defaults to \code{LengthClass}.}
}
\value{
The input table with an additional column \code{length_cm}.
}
\description{
This function adds a new column \code{length_cm} to the input table (\code{d}),
computed based on \code{LengthCode} and \code{LengthClass} (or their alternative names).
}
