% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_products.R
\name{dr_catch_weight}
\alias{dr_catch_weight}
\title{Get catch weights}
\usage{
dr_catch_weight()
}
\description{
Calculate the total reported catch numbers and weight by species and haul.
}
\details{
The functions is supposed to give the same results as icesDatras::getCatchWgt
(needs checking), except gives latin (and english) name.
}
