% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connetion_parquet.R
\name{dr_con}
\alias{dr_con}
\title{Create a DuckDB connection to a DATRAS dataset (latin names included)}
\usage{
dr_con(type = NULL, trim = TRUE)
}
\arguments{
\item{type}{A character string specifying the type of the dataset. Must be one of "HH", "HL", or "CA".}

\item{trim}{A boolean (default TRUE) which removes station variables from the HL and CA data, the join
be made via the ".id" variable.}
}
\value{
A DuckDB dataset object.
}
\description{
This function creates a URL for a DATRAS dataset based on the input \code{type} and opens the dataset
using the \code{duckdbfs::open_dataset} function. The type must be one of "HH", "HL", or "CA".
}
\details{
The connection is to an experimental table setup that includes latin name in the HL and CA files
}
\examples{
\dontrun{
  dr_con("HH")
}
}
