% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_connection.R
\name{dr_con}
\alias{dr_con}
\title{Create a DuckDB connection to a DATRAS Exchange dataset with Latin names included}
\usage{
dr_con(
  type = NULL,
  trim = TRUE,
  url = "https://heima.hafro.is/~einarhj/datras_latin/"
)
}
\arguments{
\item{type}{A character string specifying the type of dataset. Must be \code{"HH"}, \code{"HL"}, or \code{"CA"}.
This parameter maps to specific files in the provided data source.}

\item{trim}{A boolean flag (default \code{TRUE}). If \code{TRUE} and the \code{type} is \code{"HL"} or \code{"CA"},
the dataset is trimmed to ignore station-level fields.}

\item{url}{The http path to the DATRAS parquet files}
}
\value{
A DuckDB dataset object.
}
\description{
This function establishes a DuckDB connection to a DATRAS dataset based on the specified \code{type} ("HH", "HL" or "CA").
The dataset is accessed via a URL and opened using the \code{duckdbfs::open_dataset} function.
The function operates on experimental tables that include Latin names in the "HL" and "CA" files.
}
\section{Datapath Explanation}{
The function constructs a URL for the dataset based on the following pattern:
"https://heima.hafro.is/~einarhj/datras_latin/{type}.parquet".
The \code{type} parameter determines the specific file to connect to, where:
\itemize{
\item \code{"HH"} refers to haul-level data.
\item \code{"HL"} refers to catch-at-length data.
\item \code{"CA"} refers to age-based biological sampling data.
}

Optionally, for the "HL" and "CA" types, setting \code{trim = TRUE} (the default) excludes station-level variables.
This allows a narrower view (read: fewer variables) of these observations, station-level variables can be retrieved
by a join to the haul table (HH) using the \code{.id} column.
}

\examples{
\dontrun{
  dr_con("HH")              # Connect to haul-level data.
  dr_con("HL", trim=FALSE)  # Get all fields for catch-at-length data.
}
}
