% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connetion_parquet.R
\name{dr_con_exchange}
\alias{dr_con_exchange}
\title{Create a DuckDB connection to a DATRAS dataset (exchange format)}
\usage{
dr_con_exchange(type = NULL)
}
\arguments{
\item{type}{A character string specifying the type of the dataset. Must be one of "HH", "HL", or "CA".}
}
\value{
A DuckDB dataset object.
}
\description{
This function creates a URL for a DATRAS eschange dataset based on the input \code{type} and opens the dataset
using the \code{duckdbfs::open_dataset} function. The type must be one of "HH", "HL", or "CA".
}
\examples{
\dontrun{
  dr_con_exchange("HH")
}
}
