% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_csv.R
\name{dr_get}
\alias{dr_get}
\title{Download, extract, and import DATRAS Data}
\usage{
dr_get(
  recordtype,
  surveys = NULL,
  years = 1965:2030,
  quarters = 1:4,
  from = "parquet",
  quiet = TRUE
)
}
\arguments{
\item{recordtype}{A character string indicating the record type ("HH", "HL", or "CA").}

\item{surveys}{A character string specifying the survey names.}

\item{years}{An integer vector specifying year range}

\item{quarters}{An integer vector (e.g., \code{"1:4"}).}

\item{from}{A text string any of "old", "new", "parquet".}

\item{quiet}{A logical value; if \code{FALSE}, progress messages are displayed.}
}
\value{
A data frame containing the requested DATRAS data for the specified year and quarter ranges.
}
\description{
This function automates the process of downloading, unzipping, and importing DATRAS
(ICES Database of Trawl Surveys) data into R. The year and quarter parameters must be
provided as single text strings specifying ranges (e.g., \code{"2020:2025"} for years and \code{"1:4"} for quarters).
Temporary files are automatically cleaned up.
}
