% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_csv.R
\name{dr_get}
\alias{dr_get}
\title{Download, extract, and import DATRAS Data}
\usage{
dr_get(
  recordtype,
  survey,
  year = 1965:2030,
  quarter = 1:4,
  quiet = TRUE,
  how = "data.table"
)
}
\arguments{
\item{recordtype}{A character string indicating the record type ("HH", "HL", or "CA").}

\item{survey}{A single character string specifying the survey name.}

\item{year}{A single text string specifying the range of years as \code{"start:end"} (e.g., \code{"2020:2025"}).}

\item{quarter}{A single text string specifying the range of quarters as \code{"start:end"} (e.g., \code{"1:4"}).}

\item{quiet}{A logical value; if \code{FALSE}, progress messages are displayed.}

\item{how}{Text string, any of "parquet", "arrow" or "data.table"}
}
\value{
A data frame containing the requested DATRAS data for the specified year and quarter ranges.
}
\description{
This function automates the process of downloading, unzipping, and importing DATRAS
(ICES Database of Trawl Surveys) data into R. The year and quarter parameters must be
provided as single text strings specifying ranges (e.g., \code{"2020:2025"} for years and \code{"1:4"} for quarters).
Temporary files are automatically cleaned up.
}
