% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_download_data.R
\name{dr_get_data_latin}
\alias{dr_get_data_latin}
\title{Get DATRAS tables with latin names}
\usage{
dr_get_data_latin(recordtype, survey, year, quarter, quiet = TRUE)
}
\arguments{
\item{recordtype}{Any of HH, HL, CA or all}

\item{survey}{Name of the survey, e.g. "ROCKALL"}

\item{year}{A text string, like "1999:2009"}

\item{quarter}{A text string, like "1:4"}

\item{quiet}{A boolean (default TRUE) - suppress messages
'}
}
\value{
A tibble
}
\description{
Get DATRAS tables with latin names
}
