% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_data_exchange.R
\name{dr_get_datras}
\alias{dr_get_datras}
\title{Retrieve DATRAS exchange table}
\usage{
dr_get_datras(type, survey, years = NULL, quarters = NULL, quiet = TRUE)
}
\arguments{
\item{type}{A character string, the DATRAS exchange table type (e.g., "HH").}

\item{survey}{A character string, the survey acronym (e.g., \code{"NS-IBTS"}).}

\item{years}{An integer vector of years (e.g., \code{2010} or \code{2005:2010}).}

\item{quarters}{An integer vector of quarters (1, 2, 3, or 4).}

\item{quiet}{Logical; if \code{TRUE} (default), suppress messages.}
}
\value{
A tibble with exchange data or an empty tibble if no data is available.
}
\description{
This function combines the functionality of dr_get_HH, dr_get_HL, dr_get_CA and
dr_get_FL. It supports querying many surveys, years and quarters in one
function call.
}
\details{
This function is a wrapper around \code{icesDatras::getDATRAS()} that
ensures variable types are properly set.
}
