% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_settypes.R
\name{dr_settypes}
\alias{dr_settypes}
\title{Set variable types}
\usage{
dr_settypes(d)
}
\arguments{
\item{d}{A DATRAS exchange table}
}
\value{
A table of same dimention as input
}
\description{
Used when downloading raw data to ensures consistency in column types
}
\details{
The column type setting is according to \href{"www.ices.dk\/data/Documents\/DATRAS\/DATRAS_Field_descriptions_and_example_file_May2022.xlsx"}{DATRAS_Field_descriptions_and_example_file_May2022.xlsx}
with some additional guesswork for flexfile variables
}
