% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_tidy.R
\name{dr_tidyca}
\alias{dr_tidyca}
\title{Tidy DATRAS age (CA) data}
\usage{
dr_tidyca(d, trim = TRUE)
}
\arguments{
\item{d}{A duckdb-table connection or a tibble containing CA data.}

\item{trim}{Boolean (default TRUE) remove haul data}
}
\value{
A duckdb-table connection or a tibble with tidied CA data.
}
\description{
Cleans and formats age (CA) data, converting length units, setting weights, and finalizing aphia codes.
}
